Remove an unusable delete() method from the bindings.
bodhi.client.bindings.BodhiClient.delete() could not have been
working since it uses @errorhandled and does not return any
response. The errorhandled() decorator will fail when any decorated
method returns None. Since the delete() method contained a warning
that it was deprecated, it was determined that it was better to go
ahead and delete it rather than fix it.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>