Correctly calculate how many days to stable an Update has.
The Update.days_to_stable() property only calculated the number of
days to stable for updates that were already eligible to go to
stable, in which case it returned negative values only. This commit
reworks the function so that it correctly returns the number of
days until an update can be pushed to stable, or 0 if the update
can already be pushed to stable.

fixes #1305

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>