Add a --url flag to the CLI.
This commit adds a --url flag to the CLI so that it will be easier
to develop the CLI in the development environment. It also listens
for a BODHI_URL environment variable.

This commit happened to touch several functions which had no test
coverage so it also happened to raise the test coverage by 1%.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>