Initialize the bugtracker in main() instead of on import.
Before this commit, it was not possible to build the documentation
without installing Bodhi first, since importing bodhi.server would
cause the config to get loaded which would cause a exception when
it could not find the bodhi-server distribution. With this patch,
it is possible to import Bodhi's server code without installing
Bodhi.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>