Add titles to the header icons in the Update page
Previously, the icons at the top of the Update page that
provide glancable context for the update type and critial
path status did not have titles. Consequently, if the user
did not know what they meant, they were a little cryptic.
This change adds a title to each that explains the icon when
the user hovers over them.

Fixes: #1362