Optimize the ajax calls to resultsdb
previously, when presenting the automated test results,
bodhi queried every test case possible for each build
in an update. Consequently, if a large update was encountered,
this resulted in a large amount of calls to results db.
This patch utilises the results/recent/ endpoint of resultsdb
to just give us the most recent results for a build.

Additionaly, this patch re-organizes the output of the automated
test results to order them by packagename in the automatedtests
tab.

Related: #951 #983