Show openQA results in the web UI
This includes openQA results (as well as Taskotron) in the
web UI "Automated Tests" view. Yay one-liners! It does not
allow you to gate an update on openQA tests, as that code is
separate.

We may need to make this more complex if we ever start running
openQA update tests with more complex 'scenarios' (when two
jobs have the same test name but different arch, or 'flavor',
or BIOS vs. UEFI, stuff like that). But with the set of tests
we currently run, this works fine.

Signed-off-by: Adam Williamson <awilliam@redhat.com>