Make sure ``pop`` calls load the configuration
SQLAlchemy calls ``pop`` on the configuration its given which didn't
load the configuration. This commit makes sure ``pop`` behaves like
``get`` and ``__getitem__`` (in that it loads the configuration). It
also makes sure copying the dictionary triggers a configuration load,
and uses copy to provide SQLAlchemy with a configuration to mutate.

fixes #1423

Signed-off-by: Jeremy Cline <jeremy@jcline.org>