Web requests now use the correct Session object when committing.
Prior to this commit, the web request lifecycle would commit a
different database session than the rest of the code would use to
do its work.

fixes #1470
fixes #1473

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>

This commit was also partially written by Jeremy Cline, and this is
his original commit message:

Many tests need to inspect objects tied to sessions that are
automatically removed after the session is committed and removed from
the session registry. This configures the scoped session used in the
tests to not expire objects when their session is no longer active.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>