Handle openQA 'same test case, multiple flavors'
So, this is the infamous 'scenario' problem. In openQA, we run
four of the update tests - the base_ tests - twice, once from
a Server base image, once from a Workstation base image (as the
results of the tests could potentially differ). In openQA's
terms, the 'test' is the same, the 'flavor' differs.

The term 'scenario' is one openQA came up with while working
with this problem. Basically, if two tests have the same
'scenario', it means they're "the same test". openQA defines
a constant consisting of all the metadata keys that, taken
together, define a test's 'scenario', and call them the
'scenario keys'.

When we report an openQA result to ResultsDB, we basically use
the openQA test name (transmogrified a bit) as the testcase
name, so there can be multiple openQA results with the same
testcase name that are not considered to be re-runs of each
other. We take the values of all the 'scenario keys', glue
them together (with periods), and report this string as the
'scenario' in the result extradata. This was my initial attempt
to set things up so consumers (like Bodhi) could handle the
scenario problem.

Unfortunately now we come to actually use it on the consumer
end, some problems become apparent. Firstly, you can't ask
ResultsDB something like "give me the latest results for item
X by scenario"; the `/latest` endpoint code is hardcoded to
only give you "latest by testcase name". I looked at adding a
'latest by scenario' feature somehow or other, but it's rather
hard to implement given how ResultsDB actually stores its data
(the 'extradata' values are in a different table from the core
values like 'testcase name', and the way they're stored isn't
very amenable to a trivial JOIN operation here, so you rapidly
wind up in advanced-SQL-land).

Secondly, just providing an ugly 'scenario' string misses out
on answering an important part of the problem: it doesn't help
a consumer like Bodhi know how to *display* scenario-bound
results. Just having a scenario string means we either have to
teach the consumer to parse it (as this commit, reluctantly,
does) or the consumer has to just show two apparently-identical
results and let the reader try and figure it out somehow.

This commit is a fairly ugly short-term band-aid which works
for how we *currently* run openQA update tests; in practice
the only 'scenario' element that matters currently is the
'flavor', so we handle that very similarly to how we handle
'arch' for the Taskotron `dist.depcheck` results (note that this
increases the number of queries we have to run). We may start
running the tests on multiple arches at some point, but if we
do that, we can easily extend the existing 'arch' special-casing
a bit to handle it. If we start running the tests on x86_64 and
UEFI or any other openQA 'scenario' case beyond 'flavor' and
'arch', though, we'll have the problem again.

It's worth noting that the `dist.depcheck` 'arch' case already
*is* the "scenario problem"; it's just that no-one bothered
to conceptualize it before, it just got worked around. To put it
formally, the 'scenario' for a Taskotron depcheck test consists
of the testcase name ('dist.depcheck') plus the arch; for all the
other current Taskotron tests, the 'scenario' is just the test
case name. But I think it's pretty clear that we should try and
come up with a better, more generic way to do this than just
teaching Bodhi how to deal with different 'scenario' cases as
they crop up. We need to go back to the drawing board on the
'scenario' thing and come up with a better implementation that
Bodhi and other consumers can use to handle this kind of case
without special knowledge. Perhaps we can start from @kparal 's
idea of having results define their 'scenario keys', not just
provide a glommed-together scenario string...

fixes #1471

Signed-off-by: Adam Williamson <awilliam@redhat.com>