Add a ci_status support to updates
This status is used to reflect the status of the update in the CI pipeline.

The ci_status is a property of a build, it can be any of:
- None: the update was created while bodhi was not configured to account for CI
- waiting: the update was created while bodhi was configured to support CI
- ignored: a message from the CI system said this build will not be tested
- queued: a message from the CI system said that the tests for this build have
  been queued in the system
- running: a message from the CI system said that the tests for this build are
  being run
- passed: a message from the CI system said that the tests for this build have
  passed
- failed: a message from the CI system said that the tests for this build have
  failed

Then the update has its own ci_status property which gather the information from
the builds composing that update.
In addition, an update has also a ci_passed property returning a boolean on whether
the builds composing that update have passed CI or not.

This CI status is then displayed on the update page on the UI for the users to see.
The status can be linked if needed.
The header of the CI section in the update page can also be linked to an external
page providing more/generic information about the CI integration in bodhi.

Finally, an important note: enabling CI integration in bodhi will make the application
gate on CI results, so all updates that are not passing CI will not be included in
the push to the repositories.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>
Signed-off-by: Randy Barlow <randy@electronsweatshop.com>