Don't show the edit update form if a user can't edit
Previously, both users that were not logged in, and
users logged in, but not having permissions to edit
an update could view the edit update form. These
users were correctly unable to submit the form, and
buttons that linked them to the form were not shown,
but if the URL was typed in, the user could still
access the form page.

This commit changes it so users that don't have
edit privs on an update get a 403 rather than the
edit form.

Fixes #1521