Add the signing status of an update to the update page
This changes the icon on the builds tab to signify that
a build is signed to be a key rather than a lock.

Also adds the key icon to the status area in the right
bar on the Details tab of an update if the Update is
signed.

Removed the cog icon that stated that the "Build is being Signed"
as this was just shown if the build was not signed, not that
the build was in the process of being signed.

Fixes #1469