Fix Pyramid deprecation warnings
Pyramid has deprecated a number of API items in the pyramid.security
module that Bodhi was using. These are now attributes in the Pyramid
request. This patch changes authenticated_userid,unauthenticated_userid,
and effective_principals to use the corresponding attributes in the
Pyramid request object.

This also fixes a Python decprication warning the same as the ones
fixed in PR#1498