Use bleach instead of python-markdown's Safe Mode
Previously, we used the Safe Mode of python-markdown to
replace HTML tags in the markdown input from the user. It
replaced all text with a Raw HTML not allowed message.

However, the Safe Mode feature in python-markdown was deprecated,
and they recomended that Bleach be used instead.

This commit passes the markdown input from the user through
Bleach -- before it renders the markdown to html -- escaping all
HTML tags.