not show the override form if not logged in
Previously, the full edit override form was shown when
looking at an override, even if the user was not logged
in. This commit changes the override template so the
form is not shown to the non-logged in user. Instead
they get just the info about the override.

Fixes: #1541

This also adds some tests to test the above funtionality,
as well as a few extra tests on the overrides views in
html where we didnt test in the past, fixing #946