Clean up the --help output in bodhi-client
python-click, the library we use to create the CLI for bodhi-client
takes the first docblock of a method and uses it to help create the
--help docs. Previously, these docblocks contained developer docs
that we of little use to the end user. This commit splits these
docblocks into two seperate ones, the first for user docs, (which is
nommed and used by click), and the second for developer docs.

The user docs are also expanded to include documentation on the
arguments a command requires, which was not done in the past.
click does not have a way to document the arguments of a command, like
it does for options, and they suggest that it just be done in the
main docblock for a command.

fixes #1457