mash_dir and mash_stage_dir default to None.
Previously, mash_dir and mash_stage_dir defaulted to a path
relative to the source code checkout, but this causes problems
when that path does not exist. Additionally, it does not make sense
for all Bodhi servers to have these settings defined, since the web
front end does not need to mash (nor do the various CLI tools).

With this change, the setting defaults to None and will only be
validated to exist if defined in the config file.

fixes #1611

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>