Convert the 4 koji related settings from unicode to str.
If the koji krb settings are unicode objects, the krbV library will
raise a TypeError. It turns out that the koji URL is also used in
forming some of the arguments to krbV, so it also must not be a
unicode object. This commit converts the four koji settings into
strs, and adds a test to ensure they remain strs.

fixes #1624

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>