Fix the pkgtags_url default
Before the new config system, not specifying a configuration option would set it
to an empty string or None.

The pkgtags code is entirely broken, since it overwrites the updateinfo entry in
repomd.xml because modifyrpeo only takes a filename, and then update updateinfo with
that.
Before, this was not a problem since we just did not configure pkgtags_url, which
means that pagure.server.metadata's insert_pkgtags would not do anything.
With the new config code, this code now gets used, which broke all repos.

NOTE: We should make *VERY* sure there are no other such places where code was not
used because config.get(foobar) returned None, and now returns the default value,
with an incorrect default value.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Signed-off-by: Randy Barlow <bowlofeggs@fedoraproject.org>