Use datetime.utcnow() in tests instead of now().
The tests in bodhi.server.scripts.test_approve_testing were using
datetime.now() to set the date on updates, which worked fine for
systems using UTC. However, these tests fail on systems that are in
some other timezones, such as the CentOS CI systems. This commit
converts those tests to use utcnow() to be consistent with the way the
server code uses timezones.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>