Update Typeahead JS library to version 1.1.1
Updates the typeahead bundle (containing both bloodhound and
typeahead) to the corejavascript maintained fork version 1.1.1.

https://github.com/corejavascript/typeahead.js/releases/tag/v1.1.1

The maintained fork was used instead of the unmaintained upstream
as we were hitting an issue where on some occasions not all the
search results were shown.

The typeahead API changed between releases, so this also makes
our search work with the new API.