Unify the base test classes
This unifies the base test classes by making a single, effect base
class. This base class sets up the test database with nested
transactions so that it's simple to roll the changes a test makes back
at the end of the test. It also creates a /single/ TestApp for tests to
use.

Creating a TestApp leaks a substantial amount of memory for each test
case. Tests that require a different configuration should create their
own TestApp (although this is costly as it still leaks memory at this
time).

Signed-off-by: Jeremy Cline <jeremy@jcline.org>