Pass pydocstyle arguments positionally and use --add-ignore.
Tox had been using pydocstyle's --match argument to list the paths
we wanted to enforce, but --match is meant to accept a regex and
so none of Bodhi's code was being matched and it returned 0 no
matter what. This commit converts it to pass the paths to check
positionally, similar to the old test_style.py.

Furthermore, tox had been using pydocstyle's --ignore flag instead
of it's --add-ignore flag like the old test_style.py used to do.
The difference is that --ignore implies that we want all of
pydocstyle's possible tests to be run on the code except for the
given tests, where --add-ignore says we want the default tests
with exceptions. This commit switches us back to --add-ignore so
the tests use the defaults (with one exception) and pass.

fixes #1666

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>