encode value when searching by pressing enter
the search.js script has a special method that catches
if the user presses enter on the search bar without choosing
an item from the typeahead dropdown, and just searches for the
package name. WE previously escaped the search value for items
chosen in the dropdown, but missed the special case when the user
presses enter. This escapes that value, so results for packagenames
with special URI chars in them (e.g. memtest86+) now show properly
when just typing a search string and pressing enter.