gate on the test results with Greenwave
Enabling Greenwave integration in bodhi will make the application
gate on the test results, so all the updates that do not satisfy the policies
enforced by Greewave will not be included in the push to the repositories.
Note: this change will make CI integration no longer gate on CI results.