use the ilike search for search on enter in searchbar
Previously, if the user skipped the search suggestions
in the searchbar in the webui, and just pressed enter,
we returned results for whole packages only
(i.e. /updates/?packages=).

This commit changes the endpoint we point to when the
user presses enter on the search to be /updates/?search=
and now more obvious results are returned.

Fixes: #870