Add some test coverage for bodhi.client.bindings.
In order to test this module, it was necessary to also bring an
import dnf to the module level (where it should be anyway, as per
PEP-8). Technically, DNF was required before this change or users
would hit tracebacks on some functionality in the bindings. With
this change it is now impossible to import the bindings without
dnf. Thus, this change also marks dnf as a formal dependency for
the client bindings.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>