fix docstring for the email param in bodhi.client.bindings.comment
Previously, the docstring for the email parameter in
bodhi.client.bindings.comment() incorrently stated the param
controlled email notifications. The presence of an email address
in the email= parameter makes bodhi submit the comment anonomously,
rather than by a logged in user.

This commit changes the docstring to fix this.

Fixes #289