link directly to libravatar login page rather than send a POST
Previously, bodhi was sending a POST request to the libravatar
openid login page from the user profile page for the logged in
user to change their avatar. However, this was causing issues
with CSRF protection on libravatar. This commit links to the
libravatar login page rather than sending a POST request.

Fixes: #1674