Update the Greenwave JavaScript to use the new API.
The Greenwave JavaScript code needed to be updated to use the new
Greenwave API for update decisions. This commit refactors the
existing Python code to move the subject generating code into the
Update model so that it can be shared between the check_policies
script and the JavaScript, and updates the JavaScript to use a
special _json version of it.

fixes #1758

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>