Fix unusable RSS Accept-header based redirects.
While PEP-257'ing bodhi.server.services.zz_redirects, I discovered
that there were service definitions that seemed to attempt to
redirect requests with the header "Accept: application/atom+xml" to
the appropriate rss feed url for the collection. There were two
problems with this:

0) The redirect_maker function accepted a target for the collection
   type to redirect to, but ignored it and always redirected to the
   update RSS feed. Thus, requests for the users RSS feed would be
   redirected to the updates RSS feed.

1) None of the service handlers were accessible, as the other
   existing handlers were catching requests. Thus, all requests
   with the Accept header would receive a 404.

This commit removes the unusable code, and adds simple decorators
to the existing service definitions that allow users to query the
collections with the proper Accept header without a redirect. Thus,
the intended functionality is achieved.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>