Fix an unsafe default value in bodhi.server.util.call_api().
call_api() had been initializing its data parameter to {}. This is
unsafe, because if anything modifies data in the stack, the
modification will remain for the life of the process as the
default.

This commit adjusts the function to default the data parameter to
None, and to set it to {} if None is found.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>