WIP: Add 'rerun test' button
This adds a 'rerun test' button to the 'Automated Tests' tab,
which causes Bodhi to send out a fedmsg requesting that the
test system re-run the test.

Well.

That's what I WANT it to do. What it actually does is add an
API endpoint for doing that, and add an icon that does nothing
at all to the 'Automated Tests' tab. Cleanly having that icon
act as a button that sends the post request is apparently beyond
my (more or less non-existent) Javascript skills right now, so
if someone wants to show me how to do that, I'd be grateful...

Also, no tests yet, will do that soon.