CLI prints a hint about how to use koji wait-repo on overrides.
The web interface has long printed a message to users about how to
use ``koji wait-repo`` to monitor the progress of created buildroot
overrides, which was a feature gap in the CLI. This commit gives
the CLI the same ability when users create or edit an override, or
when they query for overrides and a single result is matched.

In order to accomplish this, it was necessary to be able to match
the override Build object with a release so the CLI could print the
correct tag to wait for. To accomplish this, Builds now serialize
their release_id field, and releases can now be queried with a new
ids parameter.

fixes #1376

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>