use pungi temporary directory structure
Bodhi previously had a directory structure like this:

    f26-updates-170801.1344
    f26-updates-170802.1418
    ...
    f26-updates-170810.1517
    f26-updates -> f26-updates-170810.1517

Calling pungi has its own temporary directory structure:

    f26-updates-26-20170817.0
    f26-updates-26-20170819.0
    latest-f26-updates-26 -> f26-updates-26-20170819.0

With this change we will no longer use the bodhi unique directory
structure and we'll just use the one created by pungi for us. We'll
still use the symlink that was created in the past by bodhi so we'll
end up with this in the directory as well:

    f26-updates -> f26-updates-26-20170819.0

This symlink will be controlled by bodhi and only created once the
entire bodhi process has completed. This is unlike the latest-f26-updates-26
which is created by pungi and thus exists before the rest of the bodhi
pieces have finished.

NOTE: some of the code changes look weird mainly because we now don't
      know our "unique path" until after pungi has run. We have to wait
      for it to finish and then 'detect' the temporary path so that we
      can use it in later code.