remove saving of updateinfo cache
We should be able to just pick up the updateinfo from the previous
run on the bodhi server. No reason to maintain a separate directory
for caching this stuff. We apply the same logic to 'finding the previous
runs repodata' as was done when saving the cache before.

Also now call ExtendedMetadata with mash_dir instead of self.path
because we don't know self.path at the time ExtendedMetadata is
initialized and we don't really need to know it.