Expand the groups when querying pagure
With this commit pagure is asked to provide membership information
in addition to its usual data.
This feature was added in: https://pagure.io/pagure/pull-request/2627

The members of the group with commit or admin access are then simply
added to the list of committers.

Fixes https://github.com/fedora-infra/bodhi/issues/1810

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>