Fix multiple mashes racing for the comps directory
Without this, multiple running mashes might get in each others way with the comps
repository.
If one mash is just performing the "make", and another starts with the "git pull"
phase, the "make" might hit on filees that the "git pull" is still writing out,
leading to incorrect XML files being parsed, and masher erroring out.

This patch makes the comps repos be cloned into a directory per mash.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>