Display the push to batched button for pushed updates.
The HTML template code is crazy[0,1] and its craziness was
underestimated when the batching feature was first written. It
turns out that it would only display the push to batched button for
unpushed updates.

This commit refactors the template code to be slightly less crazy
(further refactoring is certainly needed), and displays that button
on pushed updates. It also adds several test cases to assert that
the correct buttons appear at the right times.

These new tests revealed that the CI tests were leaking state, as
they would pass on their own but fail if the CI tests were run
first. Thus, this commit also refactors the CI tests to stop
leaking state by using mock.

fixes #1875
re #1887
re #1888

[0] https://github.com/fedora-infra/bodhi/issues/1887
[1] https://github.com/fedora-infra/bodhi/issues/1888

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>