Remove double documentation and commented settings from devel file.
The settings were double-documented in Bodhi which was tricky to
keep updated. This commit removes all comments and commented
settings from development.ini.example, and moves it into the devel/
folder where it more appropriately belongs. The production.ini file
is now the one source of truth for documentation about bodhi's
settings.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>