Remove an unreachable RSS type check.
The RSS handler has a series of if statements to check which type
of thing is being rendered. If none of the things match, it raised
a 404 error stating that the item cannot be rendered. This check
was not reachable by end users, since the decorator must be present
on views that support RSS for the code to be hit, and no views have
the decorator that aren't supported. It would be an error (and not
a 404) if a view had the RSS decorator without support in this
function, so it would truly be better to blow up than to hand back
a false 404. Thus, this commit removes that unreachable check.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>