Define a Compose model to replace MASHER lockfiles.
Due to
[alembic #122](https://bitbucket.org/zzzeek/alembic/issues/122/),
it was not simple to write the included database migration in
a database backend agnostic fashion. Thus, this commit also
documents that Bodhi formally requires PostgreSQL.

This commit also happens to fix an issue (#717) where users could
not comment on Updates while they were mashing. This was due to
the extremely long transactions that Bodhi used to hold open prior
to this commit, and due to Bodhi modifying the state of the Update
early on during the process. The masher now commits whenever the
Compose's state changes and no longer modifies the Updates until
the end of the process. This means that the Update does not get
locked for hours in the database any longer.

This commit also addresses an RFE (#1245) to get bodhi-push to
collate the updates by release. It goes further by also collating
them by content type and stable/testing.

fixes #717
fixes #1245
fixes #2014

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>