Run only one container build at a time in CI.
We've been trying to figure out a problem with the device mapper in
CentOS CI. We had previously tried to stagger the container builds
to see if it was a race condition and it did not work. This commit
instead tries to run only one docker build at a time.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>