Make builds.package_id non-nullable.
While writing tests for the validators, I came across some untested
code that was checking to make sure that Builds had Pacages
associated with them. Since it would be an error for a Build to be
unassociated with a Package, I opted to delete that check and
instead make the foreign key non-nullable in the database.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>