Remove an impossible validation check and its failing test.
Two commits conspired together to create a failing test, as they
were submitted in separate pull requests and each passed tests
individually. The first added a test condition for an existing
validator that tried to make sure that edited updates have releases
associated with them. The second made it impossible for updates to
be unassociated with a release. The combination caused a test to
fail because the test tried to create an update that was not
associated with a release to test the validator.

This commit removes the test and the validator it tests, since it
is no longer needed due to the database enforcing the relationship
now.

fixes #2062

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>