Configure CI to run two builds in parallel.
We had been having stability issues with Bodhi's CI, with what
seemed like a race condition. As an experiment we tried reducing
the number of parallel container builds to 1 to see if that helped.
It did seem to solve the issue, so now I would like to experiment
with setting it to two.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>