Include update severity level in updateinfo.xml
Right now, doing a `dnf list-sec` shows e.g.:

```
FEDORA-2017-fb5e227432 Unknown/Sec. glibc-2.26-20.fc27.x86_64
```

The `Unknown` in the output is due to the fact that `updateinfo.xml`
doesn't seem to include the severity level of the security update even
though the update does have it set (to "low" in this case). RHEL
repodata on the other hand does include this, and it shows up in `yum
list-sec`.

This is an important factor for users when determining when to apply the
updates. So let's include it in the XML.

Signed-off-by: Jonathan Lebon <jlebon@redhat.com>
Signed-off-by: Randy Barlow <bowlofeggs@fedoraproject.org>