Run 4 CI build jobs in parallel.
We had some issues in the past with too many concurrent CI builds
causing race conditions. We solved it by doing only one build at a
time, but of course this is quite slow. A while ago we switched to
two builds in parallel, and this has been stable. This commit is an
experiment to see if four builds is stable.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>