Refactor the composer to split pungi code into a new base class.
In order to prepare the way for Bodhi to support composing
containers we need to split pungi-specific code out of the
MasherThread base class so the general composition code can be
reused by the soon-to-be-written container composing class.

This commit does that refactoring work, and renames some uses of
"mash" to "compose". Not all uses are renamed here, as some of them
would not be backwards compatible (see #2151).

re #2028

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>