bodhi-push no longer authenticates to Koji.
Sometimes users run bodhi-push as root. Typically, the backend
fedmsg-hub process is run as the apache user. When users ran
bodhi-push as root, this would cause the Kerberos ticket cache to
become root owned with permissions of 0600, which would prevent
the composer process from authenticating to Koji, causing the
composes to fail.

This commit adjusts the buildsys code to allow callers to request
whether they want an authenticated Koji session or not, and adjusts
bodhi-push to request an unauthenticated session.

fixes #2190

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>