bodhi.server.util.memoized now properly handles non-hashable args.
memoized used to check whether args itself was a hashable
parameter, but this was incorrect for two reasons:

0) args was always a tuple since it was collected as *args, and
   tuples themselves are hashable. Thus the if block was
   unreachable.
1) tuples can contain non-hashable items, which means that the
   tuple isn't always hashable.

This led to a funny issue where the function would crash if a list
was passed to it, despite having an inline comment stating that the
code could handle having lists passed to it.

The code now attempts to hash the tuple to determine if it is
hashable, and can now handle having lists passed to it.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>