The avatar() function now properly reads the prefer_ssl setting.
bodhi.server.util.avatar() had a trailing comma on its line that
read the value of the prefer_ssl setting. As a result, the variable
it was stored in was always a 1-tuple with a bool inside it that
corresponded to prefer_ssl's value. Since a tuple of non-0 length
is "truthy", this led to prefer_ssl always effectively being True,
no matter what the setting is.

The comma is removed in this commit.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>