Don't require coverage on two methods that will soon be removed.
The update_relationship() method is only used by the unused Stacks
feature, which is planned to be removed. get_pkg_pushers() is only
used with pkgdb integration, which is also planned to be removed.

Rather than spending the time to test code that is slated for
removal, let's mark it as uncovered code and forget about it.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>