Fix two tests that cause later tests to access the network.
Two tests were turning on a config option permanently. If these
tests were run before other tests, that config change leaked into
the other tests. Python 3 runs the tests in a slightly different
order, which causes tests to fail when run in Koji with no network.
This commit mocks those config changes so the config is not altered
when they are complete.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>