Disallow network during CI tests.
It was recently discovered that four tests were accessing the
network when run in Python 3, as the tests fail when run during
Bodhi's RPM build in Koji (which disallows network access). To
prevent this in the future, this commit configures Docker to
disallow network access during the docker run.

tox had been used to build the docs, run flake8, and to run
pydocstyle prior to this commit, but tox works by creating a
virtualenv and pip installing, which requires network. Since we
weren't leveraging tox's true mission in life, this commit removes
tox in favor of simply performing these actions directly when
needed.

It was also discovered that a few dependencies were missing from
the containers and had been getting pulled in by tox or by setup.py
at test time. This commit also addresses those issues.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>