Warm up the Release._all_releases cache during web app startup.
We've had some problems with our web services flapping in
OpenShift, and it seems that there is some trouble around caching
and concurrent queries. This commit warms up the
bodhi.server.models.Release._all_releases cache in the web app's
main() function so that it does not need to wait for a request to
run the query.

It also drops the db argument from the function since it is not
necessary.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>